requirements.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==3.0.0
Flask-Caching==2.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
packaging==23.2
psycopg2-binary
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
flask-http-response==0.0.1.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
requests==2.28.2
tomli==2.0.1
urllib3==1.26.14
Werkzeug==3.0.0
zipp==3.15.0
